requirements.txt
frida==15.0.8
requests==2.26.0
cryptography==3.4.7
pandas==1.3.0
flask==2.0.1
flask_cors==3.0.10
pillow==8.3.1
openpyxl==3.0.10